Meeting Analysis/requirements.txt
python-dotenv
openai
langchain_openai
pydub
python-docx
streamlit
moviepy
langchain
langchain_community
langchain_google_genai